+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - PR #443 Fix accuracy loss issue for snmg pagerank
 - PR #444 Fix warnings in strongly connected components
 - PR #446 Fix permission for source (-x) and script (+x) files.
+- PR #448 Import filter_unreachable
 
 # cuGraph 0.8.0 (27 June 2019)